Take out entire process of run into ops
