+++ src/bin/soma/commands/run.rs
@@ -1,8 +1,8 @@
 use clap::{Arg, ArgMatches, SubCommand};
 use hyper::client::connect::Connect;
 
-use soma::error::{Error as SomaError, Result as SomaResult};
-use soma::ops::{build_image, run_container};
+use soma::error::Result as SomaResult;
+use soma::ops::run;
 use soma::{Environment, Printer};
 
 use crate::commands::{default_runtime, App, SomaCommand};
@@ -33,24 +33,10 @@ impl SomaCommand for RunCommand {
         env: Environment<impl Connect + 'static, impl Printer>,
         matches: &ArgMatches,
     ) -> SomaResult<()> {
-        let repo_name = matches.value_of("problem").unwrap();
-        let repo_index = env.data_dir().read_repo_index()?;
-        let repository = repo_index
-            .get(repo_name)
-            .ok_or(SomaError::RepositoryNotFoundError)?;
-
-        build_image(&repository, &env, repo_name)?;
-        env.printer().write_line(&format!(
-            "successfully built image for problem: '{}'",
-            &repo_name
-        ));
-
-        let mut runtime = default_runtime();
-        let container_name = run_container(&env, repo_name, &mut runtime)?;
-        env.printer().write_line(&format!(
-            "successfully started container: '{}'",
-            &container_name
-        ));
-        Ok(())
+        run(
+            &env,
+            matches.value_of("problem").unwrap(),
+            &mut default_runtime(),
+        )
     }
 }
